requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
psycopg2-binary
google-cloud-vision
google-cloud-storage
python-dotenv
pillow